Extract anonymization from models
