+++ api/maintenance/scripts/utils/anonymize_collection.js
@@ -1,26 +1,24 @@
+const readline = require("node:readline");
+const { EJSON } = require("bson");
 
-const readline = require('node:readline');
-const { EJSON } =  require('bson');
-
-MODELS_PATH = `${__dirname}/../../../src/models`
+MODULE_PATH = `${__dirname}/../../../src/anonymization`;
 
 ANONYMIZABLE_COLLECTIONS = {
-  "applications": `${MODELS_PATH}/application`,
-  "cohesioncenters": `${MODELS_PATH}/cohesionCenter`,
-  "contracts": `${MODELS_PATH}/contract`,
-  "departmentservices": `${MODELS_PATH}/departmentService`,
-  "meetingpoints": `${MODELS_PATH}/meetingPoint`,
-  "missions": `${MODELS_PATH}/mission`,
-  "missionequivalences": `${MODELS_PATH}/missionEquivalence`,
-  "referents": `${MODELS_PATH}/referent`,
-  "sessionphase1": `${MODELS_PATH}/sessionPhase1`,
-  "structures": `${MODELS_PATH}/structure`,
-  "waitinglists": `${MODELS_PATH}/waitingList`,
-  "youngs": `${MODELS_PATH}/young`,
-  "lignebuses": `${MODELS_PATH}/ligneBus`,
-  "modificationbuses": `${MODELS_PATH}/modificationBus`,
-}
-
+  applications: `${MODULE_PATH}/application`,
+  cohesioncenters: `${MODULE_PATH}/cohesionCenter`,
+  contracts: `${MODULE_PATH}/contract`,
+  departmentservices: `${MODULE_PATH}/departmentService`,
+  meetingpoints: `${MODULE_PATH}/meetingPoint`,
+  missions: `${MODULE_PATH}/mission`,
+  missionequivalences: `${MODULE_PATH}/missionEquivalence`,
+  referents: `${MODULE_PATH}/referent`,
+  sessionphase1: `${MODULE_PATH}/sessionPhase1`,
+  structures: `${MODULE_PATH}/structure`,
+  waitinglists: `${MODULE_PATH}/waitingList`,
+  youngs: `${MODULE_PATH}/young`,
+  lignebuses: `${MODULE_PATH}/PlanDeTransport/ligneBus`,
+  modificationbuses: `${MODULE_PATH}/PlanDeTransport/modificationBus`,
+};
 
 const args = process.argv.slice(2);
 if (args.length != 1) {
@@ -40,14 +38,14 @@ function anonymizeCollectionStream(anonymize) {
     const parsed = EJSON.parse(line, { relaxed });
     const anonymized = anonymize(parsed);
     const json = EJSON.stringify(anonymized, { relaxed });
-    // process a line at a time
-    console.log(json)
+
+    console.log(json);
   });
 }
 
 if (collection_name in ANONYMIZABLE_COLLECTIONS) {
-  const anonymize = require(ANONYMIZABLE_COLLECTIONS[collection_name]).anonymize
+  const anonymize = require(ANONYMIZABLE_COLLECTIONS[collection_name]);
   anonymizeCollectionStream(anonymize);
 } else {
-  process.stdin.pipe(process.stdout)
+  process.stdin.pipe(process.stdout);
 }
