+++ api/src/anonymization/contract.js
@@ -1,0 +1,40 @@
+const { generateBirthdate, generateNewPhoneNumber, starify } = require("../utils/anonymise");
+
+function anonymize(item) {
+  item.tutorFirstName && (item.tutorFirstName = starify(item.tutorFirstName));
+  item.tutorLastName && (item.tutorLastName = starify(item.tutorLastName));
+  item.tutorEmail && (item.tutorEmail = "*******@*******.***");
+  item.youngFirstName && (item.youngFirstName = starify(item.youngFirstName));
+  item.youngLastName && (item.youngLastName = starify(item.youngLastName));
+  item.youngBirthdate && (item.youngBirthdate = generateBirthdate());
+  item.youngEmail && (item.youngEmail = "*******@*******.***");
+  item.youngPhone && (item.youngPhone = generateNewPhoneNumber());
+  item.youngAddress && (item.youngAddress = starify(item.youngAddress));
+  item.parent1FirstName && (item.parent1FirstName = starify(item.parent1FirstName));
+  item.parent1LastName && (item.parent1LastName = starify(item.parent1LastName));
+  item.parent1Email && (item.parent1Email = "*******@*******.***");
+  item.parent1Adress && (item.parent1Adress = starify(item.parent1Adress));
+  item.parent1Phone && (item.parent1Phone = generateNewPhoneNumber());
+  item.parent2FirstName && (item.parent2FirstName = starify(item.parent2FirstName));
+  item.parent2LastName && (item.parent2LastName = starify(item.parent2LastName));
+  item.parent2Email && (item.parent2Email = "*******@*******.***");
+  item.parent2Adress && (item.parent2Adress = starify(item.parent2Adress));
+  item.parent2Phone && (item.parent2Phone = generateNewPhoneNumber());
+  item.missionName && (item.missionName = starify(item.missionName));
+  item.missionAdress && (item.missionAdress = starify(item.missionAdress));
+  item.missionZip && (item.missionZip = starify(item.missionZip));
+  item.missionObjective && (item.missionObjective = starify(item.missionObjective));
+  item.missionAction && (item.missionAction = starify(item.missionAction));
+  item.missionFrequence && (item.missionFrequence = starify(item.missionFrequence));
+  item.missionDuration && (item.missionDuration = starify(item.missionDuration));
+  item.projectManagerFirstName && (item.projectManagerFirstName = starify(item.projectManagerFirstName));
+  item.projectManagerLastName && (item.projectManagerLastName = starify(item.projectManagerLastName));
+  item.projectManagerEmail && (item.projectManagerEmail = "*******@*******.***");
+  item.structureName && (item.structureName = starify(item.structureName));
+  item.structureManagerEmail && (item.structureManagerEmail = "*******@*******.***");
+  item.structureManagerFirstName && (item.structureManagerFirstName = starify(item.structureManagerFirstName));
+  item.structureManagerLastName && (item.structureManagerLastName = starify(item.structureManagerLastName));
+  return item;
+}
+
+module.exports = anonymize;
