+++ api/src/anonymization/waitingList.js
@@ -1,0 +1,9 @@
+const { generateRandomEmail, generateBirthdate } = require("../utils/anonymise");
+
+function anonymize(item) {
+  item.mail && (item.mail = generateRandomEmail());
+  item.birthdateAt && (item.birthdateAt = generateBirthdate());
+  return item;
+}
+
+module.exports = anonymize;
