+++ api/src/models/PlanDeTransport/ligneBus.js
@@ -3,8 +3,9 @@ const mongooseElastic = require("@selego/mongoose-elastic");
 const esClient = require("../../es");
 const patchHistory = require("mongoose-patch-history").default;
 const { getCohortNames } = require("snu-lib");
+const anonymize = require("../../anonymization/PlanDeTransport/ligneBus");
+
 const MODELNAME = "lignebus";
-const { generateRandomName, generateBirthdate, generateRandomEmail, generateNewPhoneNumber } = require("../../utils/anonymise");
 
 const BusTeam = new mongoose.Schema({
   role: {
@@ -235,20 +236,9 @@ const Schema = new mongoose.Schema({
   deletedAt: { type: Date },
 });
 
-function anonymize(item) {
-  item.team &&
-    (item.team = item.team.map((t) => {
-      t.lastName && (t.lastName = generateRandomName());
-      t.firstName && (t.firstName = generateRandomName());
-      t.birthdate && (t.birthdate = generateBirthdate());
-      t.mail && (t.mail = generateRandomEmail());
-      t.phone && (t.phone = generateNewPhoneNumber());
-      return t;
-    }));
-  return item;
+Schema.methods.anonymise = function () {
+  return anonymize(this);
 };
-
-Schema.methods.anonymise = function() { return anonymize(this); };
 
 Schema.virtual("user").set(function (user) {
   if (user) {
@@ -281,4 +271,3 @@ Schema.index({ cohort: 1, busId: 1 }, { unique: true });
 
 const OBJ = mongoose.model(MODELNAME, Schema);
 module.exports = OBJ;
-module.exports.anonymize = anonymize;
