+++ api/src/models/structure.js
@@ -1,7 +1,7 @@
 const mongoose = require("mongoose");
 const mongooseElastic = require("@selego/mongoose-elastic");
 const patchHistory = require("mongoose-patch-history").default;
-const { generateAddress, generateRandomName, starify, generateRandomEmail, generateNewPhoneNumber } = require("../utils/anonymise");
+const anonymize = require("../anonymization/structure");
 
 const esClient = require("../es");
 
@@ -315,24 +315,9 @@ const Schema = new mongoose.Schema({
   updatedAt: { type: Date, default: Date.now },
 });
 
-function anonymize(item) {
-  item.name && (item.name = generateRandomName().toUpperCase());
-  item.siret && (item.siret = starify(item.siret));
-  item.address && (item.address = generateAddress());
-  item.website && (item.website = "https://www.google.com");
-  item.description && (item.description = starify(item.description));
-  item.twitter && (item.twitter = "www.twitter.com");
-  item.facebook && (item.facebook = "www.facebook.com");
-  item.instagram && (item.instagram = "www.instagram.com");
-  //anonymize structure manager
-  item.structureManager?.firstName && (item.structureManager.firstName = generateRandomName());
-  item.structureManager?.lastName && (item.structureManager.lastName = generateRandomName());
-  item.structureManager?.mobile && (item.structureManager.mobile = generateNewPhoneNumber());
-  item.structureManager?.email && (item.structureManager.email = generateRandomEmail());
-  return item;
+Schema.methods.anonymise = function () {
+  return anonymize(this);
 };
-
-Schema.methods.anonymise = function() { return anonymize(this); };
 
 Schema.virtual("user").set(function (user) {
   if (user) {
@@ -361,4 +346,3 @@ Schema.plugin(mongooseElastic(esClient, { selectiveIndexing: true, ignore: ["jva
 
 const OBJ = mongoose.model(MODELNAME, Schema);
 module.exports = OBJ;
-module.exports.anonymize = anonymize;
